+++ tests/test_visualize.py
@@ -1,9 +1,12 @@
 from __future__ import annotations
 
+from pathlib import Path
+
 import pandas as pd
 
 from .utils import create_confounds_tsv
 from .utils import return_bidsmreye_eyetrack_tsv
+from bidsmreye.visualize import group_report
 from bidsmreye.visualize import visualize_eye_gaze_data
 
 
@@ -18,3 +21,10 @@ def test_visualize_eye_gaze_data():
     fig = visualize_eye_gaze_data(eye_gaze_data)
 
     fig.show()
+
+
+def test_group_report():
+
+    input_dir = Path().resolve().joinpath("tests", "data", "derivatives", "bidsmreye")
+
+    group_report(input_dir)
